Do not fix l=0 in cmp.set_decoder
The cmp.set_decoder definition foxes the l field to 0 (32-bit always).
In this view , cmp(l=0) and cmpw are separate instructions, and can
not have two set_asm definitions.  In an alternative view, cmpw is an
extended mnemonic.  Mixing both views is logically inconsistent.
I slightly prefer cmpw to be an extended mnemonic, because there isn't
a definition of "cmpd" in Book 1.